+++ include/formats/SipiIOTiff.h
@@ -81,7 +81,7 @@ namespace Sipi {
         * \returns Buffer of 1-bit data (padded to bytes). NOTE: This buffer has to be deleted by the caller!
         */
         unsigned char *cvrt8BitTo1bit(const SipiImage &img, unsigned int &sll);
-        
+
     public:
         static void initLibrary(void);
        /*!
@@ -100,7 +100,7 @@ namespace Sipi {
         * \param[out] width Width of the image in pixels
         * \param[out] height Height of the image in pixels
         */
-        bool getDim(std::string filepath, int &width, int &height);
+        bool getDim(std::string filepath, size_t &width, size_t &height);
 
 
        /*!
